refactor: move PackageMetadataCollection, PackagesDownloader, and IterableStreamZipOfDirectory
